Pull `FrameProperties` into its own module.
